+++ src/containerapp/azext_containerapp/containerapp_functions_decorator.py
@@ -280,7 +280,6 @@ class ContainerAppFunctionInvocationsDecorator(ContainerAppFunctionsDecorator):
             invocation_summary_query = (
                 f"requests | extend functionNameFromCustomDimension = tostring(customDimensions['faas.name']) "
                 f"| where timestamp >= ago({timespan}) "
-                f"| where cloud_RoleName =~ '{container_app_name}' "
                 f"| where isempty(\"{revision_name}\") or cloud_RoleInstance contains '{revision_name}' "
                 f"| where operation_Name =~ '{function_name}' or functionNameFromCustomDimension =~ '{function_name}' "
                 f"| summarize SuccessCount = coalesce(countif(success == true), 0), ErrorCount = coalesce(countif(success == false), 0)"
@@ -312,11 +311,9 @@ class ContainerAppFunctionInvocationsDecorator(ContainerAppFunctionsDecorator):
             revision_name = "" if self.active_revision_mode.lower() == "single" else revision_name
 
             invocation_traces_query = (
-                f"requests | extend functionNameFromCustomDimension = tostring(customDimensions['faas.name']) "
-                f"| project timestamp, id, operation_Name, success, resultCode, duration, operation_Id, functionNameFromCustomDimension, "
-                f"cloud_RoleName, cloud_RoleInstance, invocationId=coalesce(tostring(customDimensions['InvocationId']), tostring(customDimensions['faas.invocation_id'])) "
+                f"requests | extend functionNameFromCustomDimension = tostring(customDimensions['faas.name']), "
+                f"invocationId=coalesce(tostring(customDimensions['InvocationId']), tostring(customDimensions['faas.invocation_id'])) "
                 f"| where timestamp > ago({timespan}) "
-                f"| where cloud_RoleName =~ '{container_app_name}' "
                 f"| where isempty(\"{revision_name}\") or cloud_RoleInstance contains '{revision_name}' "
                 f"| where operation_Name =~ '{function_name}' or functionNameFromCustomDimension =~ '{function_name}' "
                 f"| order by timestamp desc | take {limit} "