+++ src/managedcleanroom/azext_managedcleanroom/analytics_frontend_api/__init__.py
@@ -1,0 +1,27 @@
+# coding=utf-8
+# --------------------------------------------------------------------------
+# Copyright (c) Microsoft Corporation. All rights reserved.
+# Licensed under the MIT License. See License.txt in the project root for license information.
+# --------------------------------------------------------------------------
+# pylint: disable=wrong-import-position
+# flake8: noqa
+
+from typing import TYPE_CHECKING
+
+if TYPE_CHECKING:
+    from ._patch import *  # pylint: disable=unused-wildcard-import
+
+from ._client import AnalyticsFrontendAPI  # type: ignore
+
+try:
+    from ._patch import __all__ as _patch_all
+except ImportError:
+    _patch_all = []
+from ._patch import patch_sdk as _patch_sdk
+
+__all__ = [
+    "AnalyticsFrontendAPI",
+]
+__all__.extend([p for p in _patch_all if p not in __all__])  # pyright: ignore
+
+_patch_sdk()