+++ src/databricks/HISTORY.rst
@@ -2,6 +2,18 @@
 
 Release History
 ===============
+1.3.1
++++++
+* Upgrading to 2026-01-01 API version
+* Set default SKU to Premium for serverless workspaces
+
+1.3.0
++++++
+* Revert the change of upgrading to 2026-01-01 API version
+
+1.2.1
++++++
+*  Upgrade API version from 2025-10-01-preview to 2026-01-01
 
 1.2.0
 +++++